Fixed maze bug where the top controls did not render right.

+++ trunk/source/com/barrybecker4/common/geometry/Location.java
@@ -44,11 +44,11 @@ public class Location implements Serializable {
         return col_;
     }
 
-    public int getX() {
+    public byte getX() {
         return col_;
     }
 
-    public int getY() {
+    public byte getY() {
         return row_;
     }
